requirements.txt
absl-py==0.12.0
aiohttp==3.8.1
aiosignal==1.2.0
antlr4-python3-runtime==4.9.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.5
astunparse==1.6.3
async-timeout==4.0.2
attrs==22.1.0
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.1
cachetools==4.2.1
certifi==2020.6.20
cffi==1.15.1
chardet==3.0.4
charset-normalizer==2.1.0
click==8.1.3
cloudpickle==1.2.2
colorama==0.4.5
cycler==0.11.0
debugpy==1.6.2
decorator==5.1.1
defusedxml==0.7.1
docker-pycreds==0.4.0
einops==0.3.0
entrypoints==0.4
executing==0.9.1
fastjsonschema==2.16.1
flatbuffers==1.12
fonttools==4.34.4
frozenlist==1.3.0
fsspec==2022.5.0
future==0.18.2
gast==0.3.3
gitdb==4.0.9
GitPython==3.1.27
google-auth==1.27.1
google-auth-oauthlib==0.4.3
google-pasta==0.2.0
grpcio==1.32.0
gym==0.15.3
h5py==2.10.0
hydra-core==1.2.0
idna==2.10
importlib-resources==5.9.0
ipykernel==6.15.1
ipython==8.4.0
ipython-genutils==0.2.0
ipywidgets==7.7.1
jedi==0.18.1
Jinja2==3.1.2
joblib==0.16.0
jsonpatch==1.26
jsonpointer==2.0
jsonschema==4.8.0
jupyter==1.0.0
jupyter-client==7.3.4
jupyter-console==6.4.4
jupyter-core==4.11.1
jupyterlab-pygments==0.2.2
jupyterlab-widgets==1.1.1
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
kornia==0.6.8
lightly==1.2.35
lightly-utils==0.0.2
Markdown==3.3.4
MarkupSafe==2.1.1
matplotlib==3.5.2
matplotlib-inline==0.1.3
mistune==0.8.4
multidict==6.0.2
nbclient==0.6.6
nbconvert==6.5.0
nbformat==5.4.0
nest-asyncio==1.5.5
notebook==6.4.12
numpy==1.23.1
oauthlib==3.1.0
omegaconf==2.2.3
opencv-python==4.1.2.30
opt-einsum==3.3.0
packaging==21.3
pandas==1.5.2
pandocfilters==1.5.0
parso==0.8.3
pathtools==0.1.2
pickleshare==0.7.5
Pillow==8.1.2
prometheus-client==0.14.1
promise==2.3
prompt-toolkit==3.0.30
protobuf==3.15.5
psutil==5.9.1
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.19
pyDeprecate==0.3.2
pygame==1.9.6
pyglet==1.5.2
Pygments==2.12.0
pymunk==5.5.0
pyparsing==3.0.9
pyrsistent==0.18.1
python-abc==0.1.0
python-dateutil==2.8.2
pytorch-lightning==1.7.7
pytz==2022.6
pywin32==304
pywinpty==2.0.6
PyYAML==6.0
pyzmq==23.2.0
qtconsole==5.3.1
QtPy==2.1.0
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.7.2
scikit-learn==0.23.1
scipy==1.8.1
seaborn==0.12.1
Send2Trash==1.8.0
sentry-sdk==1.7.2
setproctitle==1.2.3
shortuuid==1.0.9
six==1.16.0
sklearn==0.0
smmap==5.0.0
soupsieve==2.3.2.post1
stack-data==0.3.0
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
termcolor==1.1.0
terminado==0.15.0
threadpoolctl==2.1.0
tinycss2==1.1.1
torch==1.9.0+cu111
torchmetrics==0.9.3
torchvision==0.10.0+cu111
tornado==6.2
tqdm==4.64.0
traitlets==5.3.0
typing_extensions==4.3.0
urllib3==1.25.11
wandb==0.12.21
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.2.1
widgetsnbextension==3.6.1
yarl==1.7.2
zipp==3.8.1
